Fixed issue where Crosshair, and Drowning Indicator stayed enabled after death.
When working this issue, I saw that there were issues with the crosshair enabling and disabling if I had it added to AliveHud. When I would respawn it wouldn't enable even when I enabled it specifically, but if I would toggle console on and off, it would show up. Maybe you all understand what happened there.

I would assume that this is because of us toggling it on and off AliveHud, so I left it in the GeneralHud even though it fits in AliveHud technically.

Removed adding children to the Hud's parent as discussed in discord.

Added fullscreen to both huds as discussed in discord.

Removed unused using directive, and constructor parameters.