First Concept
Idea: Create a system that keeps track of last player interaction, and kick players that have not interacted after x configurable time.

Original concept: Client keeps track of AFK time. They keep track of last known rotation, and when a key is pressed. Server is used by server command to kick the caller of the console command from the server.